+++ Bank/src/model/ProductVO.java
@@ -1,4 +1,6 @@
 package model;
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -10,7 +12,7 @@ public class ProductVO {
 	public final String ClassName = "PRODUCT";
 	private int productId; // 상품번호
 	private String productName; // 상품명
-	private double interestRate; // 이자율
+	private BigDecimal interestRate; // 이자율
 	private int periodMonths; // 가입기간
 
 }
